refactor: dev client, server component 분리
